+++ mtg_proxy_printer/ui/settings_window.py
@@ -133,6 +133,7 @@ class SettingsWindow(*inherits_from_ui_file_with_name("settings_window")):
     def _get_download_settings_widgets(self):
         widgets_with_settings: typing.List[typing.Tuple[QCheckBox, str]] = [
             (self.include_cards_depicting_racism, "download-cards-depicting-racism"),
+            (self.include_cards_without_images, "download-cards-without-images"),
             (self.include_white_bordered_cards, "download-white-bordered"),
             (self.include_gold_bordered_cards, "download-gold-bordered"),
             (self.include_funny_cards, "download-funny-cards"),